compiles on pc, and teensy with use of the mag function
