+++ micromouse/RobotIO.cpp
@@ -2,7 +2,7 @@
 #include "IRSensor.h"
 #include "Vector.h"
 #include "Logger.h"
-
+#include "Magnetometer.h"
 #ifdef __MK20DX256__ //this is the Teensy signature
 // ### This line causes a lot of problems. It seems to conflict with #include <Encoder.h> ###
 // ### I'm not sure what it's for, so for now it's commented out.
